Moving envoy startup code out of pilot
We have a few places in the codebase that duplicate the logic for starting up envoy instances. This is the first part of an effort to merge all of these into a common utility.  For now, this is just moving the code as-is.
