+++ pkg/envoy/testdata/bootstrap.json
@@ -1,0 +1,1 @@
+{"key": "value"}
